==> scripts/03-method_characterisation/07-quality_control.Rmd <==
---
title: "implementation quality control"
output:
  html_document:
    df_print: paged
---

```{r setup}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
library(dynbenchmark)
library(tidyverse)
library(cowplot)
```

```{r message=F}
experiment("03-method_characterisation")

implementation_qc <- readRDS(derived_file("implementation_qc.rds"))
implementations_evaluated <- readRDS(derived_file("implementations_evaluated.rds"))
checks <- readRDS(derived_file("checks.rds"))

implementation_order <- implementations_evaluated %>%
  arrange(-qc_score) %>% 
  filter(!is.na(qc_score)) %>% 
  pull(implementation_id)

label_implementation <- function(implementation_id) {implementations_evaluated$implementation_name[match(implementation_id, implementations_evaluated$implementation_id)]}
```

Calculate the overall qc scores for each implementation
```{r}
# set the ordering of implementation_ids to this ordering
implementation_ordering_plot <- implementations_evaluated %>%
  mutate(implementation_id = factor(implementation_id, implementation_order)) %>% 
  ggplot(aes(as.numeric(implementation_id), qc_score)) +
    geom_bar(aes(fill=qc_score), stat="identity") +
    geom_text(aes(y=0.025, label=label_implementation(implementation_id), color=ifelse(qc_score > 0.5, "black", "white")), angle=90, hjust=0, size=5) +
    theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1)) +
    scale_y_continuous(NULL, expand=c(0, 0), limits=c(0, 1)) +
    scale_x_continuous("", breaks=NULL, expand = c(0, 0)) +
    viridis::scale_fill_viridis(label_long("qc_score"), option="D", guide=guide_colorbar(title.position = "top", title.hjust=0.5, barwidth = unit(2, "inches"))) +
    scale_color_identity() +
    annotate("segment", x=-Inf, xend=Inf, y=-Inf, yend=-Inf)+
    theme(legend.position="top")

implementation_ordering_plot
implementation_ordering_plot %>% write_rds(figure_file("implementation_qc_ordering.rds"))
```

```{r, fig.width=10, fig.height=12}
qc_checks <- implementation_qc %>%
  filter(implementation_id == "scorpius") %>%
  group_by(category) %>% 
  mutate(normalised_weight = weight/sum(weight))

aspect_position <- qc_checks %>% 
  arrange(aspect_id) %>%
  group_by(aspect_id) %>%
  summarise(
    aspect_width = sum(item_weight * normalised_weight),
    name = first(name)
  ) %>%
  mutate(
    aspect_end = cumsum(aspect_width),
    aspect_start = lag(aspect_end, 1, default=0),
    aspect_mid = aspect_start + (aspect_end - aspect_start)/2
  )

check_position <- qc_checks %>%
  arrange(aspect_id, check_id) %>%
  group_by(check_id) %>%
  summarise(
    check_width = item_weight * normalised_weight,
    aspect_id = first(aspect_id)
  ) %>%
  left_join(aspect_position, by="aspect_id") %>%
  group_by(aspect_id) %>%
  mutate(
    check_start = aspect_start + lag(cumsum(check_width), default=0),
    check_end = aspect_start + cumsum(check_width)
  )

hline_check <- geom_hline(aes(yintercept=check_end), data=check_position, color="#444444", alpha=0.25)
hline_aspect <- geom_hline(aes(yintercept=aspect_end), data=aspect_position, color="#444444")

implementation_qc_checks <- implementation_qc %>% 
  mutate(implementation_id = factor(implementation_id, implementation_order)) %>% 
  left_join(check_position, by=c("check_id", "name", "aspect_id")) %>%
  mutate(answer = ifelse(is.na(answer), 0, answer))

check_heatmap_plot <- implementation_qc_checks %>% 
  ggplot() +
    geom_rect(aes(ymax=check_start, ymin=check_end, xmin=as.integer(implementation_id), xmax=as.integer(implementation_id) + 1, fill=category, alpha=answer)) +
  hline_check + hline_aspect +
    geom_vline(aes(xintercept = as.integer(implementation_id)), color="white", alpha=0.2) +
    scale_y_reverse(NULL, breaks = aspect_position$aspect_mid, labels = aspect_position$name, expand = c(0, 0)) +
    # scale_x_continuous("", breaks = seq_along(levels(implementation_qc_checks$implementation_id))+0.5, labels = label_implementation(levels(implementation_qc_checks$implementation_id)), expand = c(0, 0), position="right") +
    scale_x_continuous("", breaks=NULL, expand = c(0, 0)) +
    scale_fill_manual("Category", values=set_names(qc_categories$color, qc_categories$category), labels=set_names(qc_categories$label, qc_categories$category), guide=guide_legend(ncol=2)) +
    scale_alpha_continuous("Score", guide=guide_legend(ncol=5, label.position = "bottom")) +
    theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1), axis.text.y = element_text(size=12), legend.position="bottom")

check_heatmap_plot
```

```{r}
check_scores <- implementation_qc_checks %>% group_by(check_id) %>% summarise(score = mean(answer))
aspect_scores <- implementation_qc_checks %>% group_by(aspect_id) %>% summarise(score = mean(answer))

check_difficulty_data <- checks %>% 
  left_join(check_scores, "check_id") %>% 
  left_join(check_position, "check_id")

check_difficulty_heatmap <- check_difficulty_data %>% 
  ggplot() + 
    geom_rect(aes(ymin=check_start, ymax=check_end, xmin=0, xmax=1), fill="#DDDDDD") +
    geom_rect(aes(ymin=check_start, ymax=check_end, xmin=0, xmax=1, fill=score)) +
    ggrepel::geom_label_repel(
      aes(
        1, 
        (check_end+check_start)/2,
        label=label_wrap(item, 35)
      ), 
      nudge_x = 0.5, 
      direction="y", 
      data=check_difficulty_data %>% 
        filter(score <= 0.5), 
      size=5, 
      lineheight=0.8, 
      min.segment.length = 0,
      hjust=0
    ) +
    geom_segment(aes(x=0, xend=1, y=check_end, yend=check_end), data=check_position, color="#444444", alpha=0.25) +
    geom_segment(aes(x=0, xend=1, y=aspect_end, yend=aspect_end), data=aspect_position, color="#444444") +
    # hline_check + hline_aspect +
    # scale_fill_manual(values=set_names(qc_categories$color, qc_categories$category)) +
    # scale_fill_distiller("QC\ndifficulty", palette="Spectral", limits=c(0,1), breaks=c(0,0.5, 1), direction=1) +
    scale_fill_gradientn("Average QC score across methods", colors = c("#9e0142", "#d53e4f", "#f46d43", "#fee08b", "#abdda4", "#1a9850"), limits=c(0,1), breaks=c(0,0.5, 1), guide=guide_colorbar(title.position = "bottom", title.hjust=0.5, barwidth = unit(2, "inches"))) +
    # viridis::scale_fill_viridis("QC\ndifficulty", limits=c(0,1), breaks=c(0,0.5, 1)) +
    scale_x_continuous(NULL, breaks=NULL, limits=c(0, 12), expand=c(0,0)) +
    scale_y_reverse(NULL, breaks=NULL, expand=c(0,0)) +
    annotate("segment", x=-Inf, xend=-Inf, y=-Inf, yend=Inf)+
    theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1), legend.position = "bottom")
  
check_difficulty_heatmap
```


Combined plot
```{r, fig.width=10, fig.height=12}
empty_x <- theme(
  axis.title.x=element_blank(),
  axis.text.x=element_blank(),
  axis.ticks.x=element_blank()
)
empty_y <- theme(
  axis.title.y=element_blank(),
  axis.text.y=element_blank(),
  axis.ticks.y=element_blank(),
  axis.line.y = element_blank()
)
empty_margin <- theme(plot.margin = margin(0.05, 0.05, 0.05, 0.05, unit="inches"))

a <- cowplot::plot_grid(
  implementation_ordering_plot + empty_margin,
  NULL,
  check_heatmap_plot + empty_margin, 
  NULL,
  nrow=2, ncol=2,
  align = "v", 
  axis = "rl", 
  rel_heights = c(0.2, 0.8),
  rel_widths = c(0.8, 0.3)#,
  # labels = c("a", "", "b", "c")
)
a
b <- cowplot::plot_grid(
  NULL,
  NULL,
  check_heatmap_plot + empty_margin, 
  check_difficulty_heatmap +empty_y + empty_margin,
  nrow=2, ncol=2,
  align = "h", 
  axis = "bt", 
  rel_heights = c(0.2, 0.8),
  rel_widths = c(0.8, 0.3)#,
  # labels = c("a", "", "b", "c")
)
b

g <- ggdraw(b)+draw_plot(a)

g

g %>% save_plot(figure_file("implementation_qc_overview.svg"), ., base_width=15, base_height=18)
```


Table for the different checks
```{r, results="asis"}
library(kableExtra)

invisible_grouper <- c(latex="\\iffalse {label_long(aspect_id)} \\fi {.}", html = "<span style='display: none;'>{aspect_id}</span> {.}")


citation_format <- c(
  latex=function(references) {
    if(is.na(references)) {
      ""
    } else {
      references %>% 
        gsub("@", "", .) %>% 
        str_split("[, \\n]+") %>% 
        first() %>% 
        glue::collapse(",") %>% 
        paste0("\\cite{", ., "}")
    }
  },
  html=function(references) {
    if(is.na(references)) {
      ""
    } else {
      references %>% 
        str_split("[, \\n]+") %>% 
        first() %>% 
        glue::collapse(";") %>% 
        {pritt("[{.}]")}
    }
  }
)

collapser <- c(latex="\\newline ", html="\n")

table <- map(c("latex", "html"), function(format) {
  table <- checks %>% 
    mutate(
      weight = cell_spec(
        weight %>% {pritt(invisible_grouper[[format]])}, # add an invisible grouping indicator
        format,
        color = spec_color(weight,option="viridis",direction=-1),
        escape=FALSE
      )
    ) %>%
    mutate(
      references = map_chr(references, citation_format[[format]])
    ) %>% 
    mutate(
      category = cell_spec(
        label_long(category),
        format,
        color = set_names(qc_categories$color, qc_categories$category)[category],
        escape=FALSE
      )
    ) %>% 
    select(category, aspect, weight, references, item, item_weight) %>% 
    rename_all(label_long) %>% 
    knitr::kable(format, escape=FALSE) %>% 
    kable_styling(bootstrap_options = "striped", latex_options = c("scale_down")) %>% 
    landscape() %>% 
    collapse_rows(1:4)

  # if (format == "html") {
  #   for (category in unique(checks$category)) {
  #     color <- qc_categories %>% filter(category == !!category) %>% pull(color)
  #     table <- table %>%
  #       group_rows(
  #         label_long(category),
  #         min(which(checks$category == category)),
  #         max(which(checks$category == category)),
  #         label_row_css = pritt("background-color: {color}; color: #fff;")
  #     )
  #   }
  # }

  
  table
}) %>% set_names(c("latex", "html"))
table
table %>% write_rds(figure_file("qc_checks.rds"))
```


Comparing qc and citations overtime     
```{r, fig.width=16, fig.height=8}
g1 <- ggplot(implementations_evaluated, aes(date, qc_score)) +
    geom_smooth(span=2) +
    geom_point() +
    ggrepel::geom_text_repel(aes(label = label_implementation(implementation_id))) +
    labs(x = label_long("publication_date"), y = "QC score", title = "Code quality score over time")
g1

g2 <- ggplot(implementations_evaluated %>% mutate(ncitations = ifelse(ncitations == 0, 0.5, ncitations)), aes(date, ncitations)) +
    geom_smooth(span=2) +
    geom_point() +
    ggrepel::geom_text_repel(aes(label = label_implementation(implementation_id))) +
    scale_y_log10() +
    labs(x = "Time", y = "Citations", title = "Citations over time")
g2

g3 <- ggplot(implementations_evaluated %>% mutate(ncitations = ifelse(ncitations == 0, 0.5, ncitations)), aes(ncitations, qc_score)) +
    geom_smooth(span=2) +
    geom_point() +
    ggrepel::geom_text_repel(aes(label = label_implementation(implementation_id))) +
    scale_x_log10() +
    labs(x = "Citations", y = "QC score", title = "QC score over # citations")
g3

cowplot::plot_grid(g1, g3, g2, ncol=2, nrow=2, labels = "auto") %>% 
  save_plot(figure_file("qc_over_time.svg"), ., base_width=10, base_height=10)
```

## Order checks

Order by breakpoint (when ordering all implementations by score)
```{r}
breakpoint <- function(x) {
  map_dbl(0:length(x), function(i) {
    sum(1-x[seq(1, i, length.out=i)]) +
      sum(x[seq(i, length(x), length.out=length(x)-i)])
  }) %>% which.min()
}
check_breakpoints <- implementation_qc %>%
  reshape2::acast(check_id~implementation_id, value.var="answer") %>%
  {
    tibble(
      breakpoint=apply(., 1, breakpoint),
      check_id = as.integer(rownames(.)),
      n_bad_implementations = apply(., 1, sum)
    )
  } %>% 
  left_join(checks, by="check_id")
```

```{r fig.width=20}
implementation_qc %>% 
  mutate(answer = round(answer, 2)) %>% 
  group_by(check_id) %>% 
  count(answer) %>% 
  mutate(check_score = sum(answer * n)) %>% 
  ungroup() %>% 
  left_join(checks, by="check_id") %>% 
  arrange(-check_score) %>% 
  mutate(check_id = factor(check_id, levels=unique(check_id))) %>% 
  ggplot() +
    geom_bar(aes(check_id, n, fill=answer), stat="identity") +
    coord_flip() +
    scale_y_reverse(expand=c(0, 0)) +
    scale_x_discrete(labels=setNames(implementation_qc$item_weight, implementation_qc$check_id)) +
    scale_fill_distiller(palette="Blues", direction = 1)
```

```{r}
aspect_info <- implementation_qc %>% 
  group_by(aspect_id) %>% 
  summarise(
    name = first(name),
    category = first(category)
  )

aspect_ordering_plotdata <- implementation_qc %>%
  group_by(implementation_id, aspect_id) %>%
  summarise(
    qc_score=sum(answer * item_weight * weight)/sum(item_weight * weight)
  ) %>%
  ungroup() %>% 
  group_by(aspect_id) %>% 
  count(qc_score) %>% 
  mutate(aspect_ordering = sum((qc_score < 1) * n)) %>% 
  ungroup() %>% 
  arrange(aspect_ordering) %>% 
  left_join(aspect_info, by="aspect_id") %>% 
  mutate(aspect_id = factor(aspect_id, levels=unique(aspect_id)))

aspect_ordering_plotdata2 <- aspect_ordering_plotdata %>% 
  mutate(category = factor(category, levels=qc_categories$category)) %>% 
  arrange(-as.numeric(category), aspect_ordering) %>% 
  mutate(aspect_id = factor(aspect_id, levels=as.character(unique(aspect_id))))

aspect_ordering_plotdata2_category <- aspect_ordering_plotdata2 %>% 
  group_by(category, aspect_id) %>% 
  summarise() %>% 
  ungroup() %>% 
  mutate(position = n()-seq_len(n())) %>% 
  group_by(category) %>% 
  summarise(start = min(position)+0.5, end=max(position)+1.5)

aspect_ordering_plot2 <- aspect_ordering_plotdata2 %>% 
  arrange(qc_score) %>% 
  ggplot() +
    geom_bar(aes(aspect_id, n, fill=set_names(qc_categories$color, qc_categories$category)[category], alpha=qc_score), stat="identity") +
    geom_vline(aes(xintercept=start), data=aspect_ordering_plotdata2_category) +
    scale_fill_identity() +
    scale_alpha_continuous(range=c(1, 0.05), name = "Quality control score", guide = guide_legend(title.position = "top", title.hjust = 0.5,label.position = "bottom"))
aspect_ordering_plot2

category_legend <- aspect_ordering_plotdata2 %>% 
  group_by(category, aspect_id) %>% 
  summarise() %>% 
  ungroup() %>% 
  mutate(position = seq_len(n())) %>% 
  group_by(category) %>% 
  summarise(start = -min(position), end=-max(position)-1) %>% 
  ggplot() +
    geom_rect(aes(xmin=0, xmax=1, ymin=start, ymax=end), fill="white") +
    geom_text(aes(1, start + (end-start)/2, label=label_long(category), color=category), hjust=1, size=5) +
    geom_segment(aes(x=1.1, xend=1.1, y=start-0.2, yend=end+0.2, color=category)) +
    scale_color_manual(values=set_names(qc_categories$color, qc_categories$category)) +
    scale_y_continuous(expand=c(0, 0)) +
    theme_void() +
    theme(legend.position="none")
category_legend

n_implementations <- length(unique(implementation_qc$implementation_id))

aspect_ordering_plot2 <- aspect_ordering_plot2 +
    coord_flip() +
    scale_y_continuous(expand=c(0, 0), name = "# implementations", c(seq(0, n_implementations, 5), n_implementations)) +
    scale_x_discrete(labels=setNames(aspect_ordering_plotdata$name, aspect_ordering_plotdata$aspect_id), name="") +
    theme(legend.position="top")

implementation_qc_ordering_plot <- cowplot::plot_grid(category_legend, aspect_ordering_plot2, align = "h", rel_widths = c(0.2, 0.8), axis="tb")
implementation_qc_ordering_plot
saveRDS(implementation_qc_ordering_plot, figure_file("implementation_qc_issues.rds"))
```
